requirements.txt
requests
xmltodict
pandas
numpy
python-dotenv